Added the ability for orange to test programs.
Like the original river, Orange will recursively go through folders in the test/ folder, testing all files. However, if a folder contains an orange.settings.json, it will run that folder as a project.

The test functionality has been improved from river: the average run time between all tests is included, along with the test that took the longest to run.

Another difference is that only the file names show as errors, not the full paths, as it was in river.