Added new array test.
arrays_morphing_future.or will test to see if the type of an array changes when you reassign one of its members to a different type.

For example,
	a = [0, 1]
is an array of int.

However,
	a = [0, 1]
	a[0] += 1.0
starts as an array of int, but morphs to an array of doubles once the analysis pass goes over the second line.