Update named node matching mechanism
Instead of passing in a list of type hints (like one for each parameter), just pass in one (the full function type). Define that VarTypes can be used as wildcards for type comparison.

Fix a problem in ASTNode::findNamed where there may have been a matching named node in a higher scope, but an exception would've been thrown on the lower-level scope.