Fix Function::matchesType.
If a function is an instance, it should only check for matching types when the parent's type was a var. The reason for this is that non-generic types in the parent may be casted to by the caller, and a perfect matching instance will never be made.

If type matching is required on those other arguments, it is handled when matchesType is called on the generic parent.