Add header files for classes
Most of them are empty at the moment, but these should be all the classes necessary to implement classes.

`ClassDecl` represents the actual class declaration.
`ClassValue` is an instance of a ClassDecl
`ClassMethod` is a method of a class.
`ClassTopLevel` is an interface to say that a node can be an immediate child of ClassDecl.
`ExprFactory` is an interface to be used for creating expressions given a hint.
`HintProvider` provides a hint. It will be detected by ASTNode::findNamed and passed to Block::findNamed.
`MemberVarDecl` is a variable declaration of a class member.