Add Type::copyType.
Gets a copy of a type, if applicable. For example, if a type is an aggregate type (i.e., contains another type inside of its definition), it will get a new instance of the type with a copy of that internal type.

The internal type's copy will only be different if it is a ReferenceType.