Change parse functions to methods of a parser
This allows us to store the lexeme stream inside the parser as a state, and allows us to expand the parser in the future to have more state.

Particularly, this will be helpful for generic instantiation (e.g., `a<int>()`) so we can keep track of when we’re in the middle of parsing a generic argument list.